Moved generator functions to own code file
