+++ test/TestHelper.h
@@ -5,51 +5,15 @@
 #include <gtest/gtest.h>
 #include <vector>
 
-#include "BusBuffer.h"
 #include "CoSimTypes.h"
-#include "Communication.h"
 
 namespace DsVeosCoSim {
 
-inline static std::string g_TestServerName = "TheTestServer";
-
-int Random(int min, int max);
-
-void FillWithRandom(uint8_t* data, size_t length);
-
-template <typename T>
-T GenerateRandom(T min, T max) {
-    return (T)Random((int)min, (int)max);
-}
-
 void AssertByteArray(const uint8_t* expected, const uint8_t* actual, size_t size);
-
-uint8_t GenerateU8();
-uint16_t GenerateU16();
-uint32_t GenerateU32();
-uint64_t GenerateU64();
-int64_t GenerateI64();
-std::string GenerateString(const std::string& prefix);
 
 void OnLogCallback(Severity severity, std::string_view message);
 void AssertLastMessage(std::string_view message);
 void ClearLastMessage();
-
-void CreateSignal(IoSignalContainer& container);
-
-void CreateController(CanControllerContainer& container);
-void CreateController(EthControllerContainer& container);
-void CreateController(LinControllerContainer& container);
-
-std::vector<IoSignalContainer> CreateSignals(uint32_t count);
-
-std::vector<CanControllerContainer> CreateCanControllers(uint32_t count);
-std::vector<EthControllerContainer> CreateEthControllers(uint32_t count);
-std::vector<LinControllerContainer> CreateLinControllers(uint32_t count);
-
-void CreateMessage(uint32_t controllerId, CanMessageContainer& container);
-void CreateMessage(uint32_t controllerId, EthMessageContainer& container);
-void CreateMessage(uint32_t controllerId, LinMessageContainer& container);
 
 void AssertEq(const IoSignal& expected, const IoSignal& actual);
 void AssertEq(const IoSignalContainer& expected, const IoSignalContainer& actual);
